Progress converting to informed, add inputs style file
