+++ package.json
@@ -43,6 +43,7 @@
     "firebase": "6.3.1",
     "history": "4.9.0",
     "hoist-non-react-statics": "3.3.0",
+    "informed": "^2.11.8",
     "js-cookie": "2.2.0",
     "lodash": "4.17.15",
     "react": "16.8.6",
