+++ src/shared/components/AppUIKit/TextField/TextField.js
@@ -1,15 +1,17 @@
 import React, { Component } from 'react'
+import { asField } from 'informed'
 
 import * as S from './styles'
 
+@asField
 export default class AppTextInput extends Component {
   render () {
-    const { label, helperText } = this.props
+    const { label, error } = this.props
 
     return <S.InputWrap>
       { label && <S.Label>{label}</S.Label> }
-      <S.TextInput {...this.props} />
-      <S.Error>{helperText}</S.Error>
+      <S.TextInput validateOnBlur {...this.props} />
+      <S.Error>{error}</S.Error>
     </S.InputWrap>
   }
 }
