+++ src/shared/components/modals/ModalController/ModalController.js
@@ -51,7 +51,7 @@ export default class ModalController extends Component {
       closeTimeoutMS: 250,
       underlayClass,
       titleText: 'my modal',
-      focusModal: true, // TODO: fix auto focus
+      focusDialog: true, // TODO: fix auto focus
       onEnter: this.onModalEnter
     }
