+++ src/shared/styles/Colors.js
@@ -1,11 +1,17 @@
-export const brand = '#E5A840'
+export const brand = '#4D8CA3'
 
 export const black = '#18171C'
+export const lightBackground = '#2E2C36'
+export const lightBackgroundActive = '#43404F'
 export const white = '#F8F8F8'
 export const transparent = 'transparent'
 export const darkPurple = 'rgb(67, 50, 72)'
+export const darkPurpleHover = '#62416C'
 
 export const black012 = 'rgba(0, 0, 0, 0.12)'
 
 export const darkGray = '#101111'
 export const adminNavBorder = 'rgba(255, 255, 255, 0.18)'
+
+// Contextual colors
+export const errorRed = '#BE4949'
