+++ src/shared/styles/index.js
@@ -7,3 +7,4 @@ export * as Shadows from './Shadows'
 export * as Sizes from './Sizes'
 export * as Admin from './Admin'
 export * as Global from './Global'
+export * as Inputs from './Inputs'
